move set-path-utils in sets-utils folder
