+++ packages/utils/other-utils/src/index.js
@@ -10,4 +10,3 @@ export {
   getCellColors,
 } from './interpolate-colors';
 export { default as Pool } from './Pool';
-export { filterPathsByExpansionAndSelection, findChangedHierarchy } from './set-path-utils';
